Prevent Bug in Migration
Tough it should not happen we check for a smaller to_date. When getting the from_date and to_date from the schedule_object we already have dates spanning over the year change - e.g. 1.12.2020 - 1.2.2021. To be sure, we are adding 1 year to the to_date if it is smaller than the from_date.