Check externalid of imported measure
If the measure was created initially in dataCycle and sent to the GIP via the GIP-Communicator, its external_key is set to its id, which is then sent by the GIP as property 'externalid'. At the first reimport the external_key is now set to the real external_key of the measure to ensure the importer will find it.